+++ addon/components/power-select-multiple/trigger.js
@@ -49,9 +49,9 @@ export default Ember.Component.extend({
   triggerMultipleInputStyle: computed('searchText.length', 'selected.length', function() {
     run.scheduleOnce('afterRender', this.get('select.actions.reposition'));
     if (this.get('selected.length') === 0) {
-      return htmlSafe('min-width: 100%;');
+      return htmlSafe('width: 100%;');
     } else {
-      return htmlSafe(`min-width: ${(this.get('searchText.length') || 0) * 0.5 + 0.5}em`);
+      return htmlSafe(`width: ${(this.get('searchText.length') || 0) * 0.5 + 0.5}em`);
     }
   }),