+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Master
 
+# 0.10.0-beta.7
 - [FEATURE] The `selected` option can now also be a promise. When it's a promise, the component
   won't have a have any selection (the trigger will be empty, no option of the list will be
   highlighted) until trat promise resolves. Once it resolves, the triger and the highlighted